Division_Macro/requirements.txt
vgamepad
pyautogui
mss
numpy
opencv-python
keyboard
win10toast-click